requirements.txt
absl-py==1.4.0
altair==4.2.2
astunparse==1.6.3
attrs==22.2.0
autopep8==2.0.1
blinker==1.5
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.4
firebase-admin==6.1.0
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
gcloud==0.18.3
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h5py==3.8.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.4.2
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
msgpack==1.0.4
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
proto-plus==1.22.2
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
Pyrebase4==4.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.1
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.18.1
streamlit-option-menu==0.3.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.13.0
